PMM, Generalized Louvain and LART
These are the versions of the algorithms that will be evaluated in the thesis. All algorithms implemented have been attributed to their respective authors and all are guaranteed to run on networks of at least size 5k, 3+ layers within 10 minutes. As always improvement points are possible in all 3 algorithms and these points will be discussed in the evaluation.

The algorithms have dtrace support (tutorial will be available in another commit) where running time and memory consumption can be measured for the entire algorithm and specific functions.

The Eigen C++ Matrix library (eigen.tuxfamily.org/) is required for all the algorithms, which is not in this commit.
In addition, the extended version of dlib and the top eigenvalue calculations library (spectralib.org/) are needed for LART and PMM, respectively. These are included in the .gitmodules file.